convert health and damages int to float

+++ Assets/Scripts/Arena/BattleSystem/Article.cs
@@ -10,7 +10,7 @@ public class Article : Target
 
     }
 
-    public override void Damage(int AttackDamage)
+    public override void Damage(float AttackDamage)
     {
         
     }
